Add base for material based rendering

+++ src/math/ray3.c
@@ -7,7 +7,7 @@
 
 #include "headers.h"
 
-dvec3 ray3_compute(ray3 ray, float t)
+dvec3 ray3_compute(ray3 ray, double t)
 {
     return (dvec3){ray.p.x + ray.v.x * t, ray.p.y + ray.v.y * t,
     ray.p.z + ray.v.z * t};
