+++ src/math/vec3_linear.c
@@ -7,17 +7,17 @@
 
 #include "headers.h"
 
-float dvec3_dot(dvec3 a, dvec3 b)
+double dvec3_dot(dvec3 a, dvec3 b)
 {
     return a.x * b.x + a.y * b.y + a.z * b.z;
 }
 
-float dvec3_norm(dvec3 vec)
+double dvec3_norm(dvec3 vec)
 {
     return sqrtf((vec.x * vec.x) + (vec.y * vec.y) + (vec.z * vec.z));
 }
 
-float dvec3_dist(dvec3 a, dvec3 b)
+double dvec3_dist(dvec3 a, dvec3 b)
 {
     return dvec3_norm(dvec3_sub(a, b));
 }
@@ -30,7 +30,7 @@ dvec3 dvec3_cross(dvec3 a, dvec3 b)
 
 dvec3 dvec3_normalize(dvec3 vec)
 {
-    float dist = dvec3_norm(vec);
+    double dist = dvec3_norm(vec);
 
     if (dist == 0.0)
         return (dvec3){0.0, 0.0, 0.0};
