+++ src/math/vec3_s.c
@@ -7,7 +7,7 @@
 
 #include "headers.h"
 
-dvec3 dvec3_adds(dvec3 vec, float value)
+dvec3 dvec3_adds(dvec3 vec, double value)
 {
     return (dvec3){vec.x + value, vec.y + value, vec.z + value};
 }
