+++ src/math/vec4.c
@@ -17,7 +17,7 @@ vec4 vec4_sub(vec4 a, vec4 b)
     return (vec4){a.x - b.x, a.y - b.y, a.z - b.z, a.w - b.w};
 }
 
-vec4 vec4_muls(vec4 vec, float value)
+vec4 vec4_muls(vec4 vec, double value)
 {
     return (vec4){vec.x * value, vec.y * value, vec.z * value, vec.w * value};
 }
